agent/requirements.txt
livekit-agents
livekit-plugins-cartesia
livekit-plugins-deepgram
google-generativeai
python-dotenv
colorama

